moved delete trade route test
